requirements.txt
pywayland>=0.4.14
pycairo>=1.23.0
dbus-python>=1.3.2